wpa_supplicant: support Multi-AP backhaul STA onboarding
The Wi-Fi Alliance Multi-AP Specification v1.0 allows onboarding of a
backhaul STA through WPS. To enable this, the backhaul STA needs to add
a Multi-AP IE to the WFA vendor extension element in the WSC M1 message
that indicates it supports the Multi-AP backhaul STA role. The registrar
(if it support Multi-AP onboarding) will respond to that with a WSC M8
message that also contains the Multi-AP IE, and that contains the
credentials for the backhaul SSID (which may be different from the SSID
on which WPS is performed).

Introduce a new, global configuration option multi_ap_backhaul_sta to
enable this feature. It has to be separate from the per-SSID
multi_ap_backhaul_sta option since no SSID exists yet when WPS is
performed. multi_ap_backhaul_sta is set to 1 in the automatically
created SSID. Thus, if the AP does not support Multi-AP, association
will fail and WPS will be terminated.

Only wps_pbc is supported.

Extend wps_build_wfa_ext() with an extra argument to add the additional
subelement to a WPS message, and use this when building M1. Note that we
don't reuse the existing MULTI_AP_SUB_ELEM_TYPE definition here, but
rather define a new WFA_ELEM_MULTI_AP, to make sure the enum of WFA
subelement types remains complete.

Signed-off-by: Davina Lu <ylu@quantenna.com>
Signed-off-by: Igor Mitsyanko <igor.mitsyanko.os@quantenna.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>