wpa_supplicant: support Multi-AP backhaul STA onboarding
The Wi-Fi Alliance Multi-AP Specification v1.0 allows onboarding of a
backhaul STA through WPS. To enable this, the backhaul STA needs to add
a Multi-AP IE to the WFA vendor extension element in the WSC M1 message
that indicates it supports the Multi-AP backhaul STA role. The registrar
(if it support Multi-AP onboarding) will respond to that with a WSC M8
message that also contains the Multi-AP IE, and that contains the
credentials for the backhaul SSID (which may be different from the SSID
on which WPS is performed).

Introduce a new, global configuration option multi_ap_backhaul_sta to
enable this feature. It has to be separate from the per-SSID multi_ap
option since no SSID exists yet when WPS is performed. multi_ap is set
to 1 in the automatically created SSID. Thus, if the AP does not support
Multi-AP, association will fail and WPS will be terminated.

Only wps_pbc is supported.

Extend wps_build_wfa_ext() with an extra argument to add the additional
subelement to a WPS message, and use this when building M1. Note that we
don't reuse the existing MULTI_AP_SUB_ELEM_TYPE definition here, but
rather define a new WFA_ELEM_MULTI_AP, to make sure the enum of WFA
subelement types remains complete.

Signed-off-by: Davina Lu <ylu@quantenna.com>
Signed-off-by: Igor Mitsyanko <igor.mitsyanko.os@quantenna.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
---
v2 (Arnout):
- Use Multi-AP instead of EasyMesh everywhere.
- Build extra WFA vendor extension subelement as part of
  wps_build_wfa_ext() instead of creating a new function
  wps_build_vendor_ext_m1_wfa().
- Inline wpas_wps_set_multi_ap_backhaul_sta() (it is just 1 line and
  used only once).
- Move multi_ap_backhaul_sta member from wps_dev to wps_context (it has
  nothing to do with the device).
- Protect all the new code with #ifdef CONFIG_MULTI_AP.
- Enable multi_ap in the new SSID configuration.

---
If we want to support the case where the received credentials are *not*
for a backhaul link, we should introduce an extra value for the
wpa_supplicant multi_ap option that indicates that it is *maybe* a
backhaul link. Then, instead of failing the association, it would set
4addr mode depending on whether or not the AP indicates it is a backhaul
link. I have not implemented it like that because typically this is not
what we want: if we expect the link to be a backhaul link and it is not,
we really want it to fail. For example, if it is a backhaul link we
would typically turn off firewalling and use a bridge instead of
IP-routing and NAT, but if it is not a backhaul link, we want to treat
it as a WAN link with firewalling and NAT. So some kind of higher level
involvement is needed.