hostapd: Add Multi-AP protocol support
The purpose of Multi-AP specification is to enable inter-operability
across Wi-Fi access points (APs) from different vendors.

This patch introduces one new configuration parameter 'multi_ap' to
enable Multi-AP functionality and to configure the BSS as a backhaul
and/or fronthaul BSS.

Advertise vendor specific Multi-AP capabilities in (Re)Association
Response frame, if the user enables Multi-AP functionality through
a configuration parameter and if the STA advertises Multi-AP in the
(Re)Assocation Request.

Validate the received Multi-AP IE. If it is a backhaul link and the
IE is provided, the station is flagged as a backhaul STA. If it is a
fronthaul link only and the IE is provided, it is an error. If it is a
backhaul link only and the IE is provided, it is an error.

A backhaul AP must support receiving both 3addr and 4addr frames from a
backhaul STA, so create a VLAN for it just like is done for WDS, i.e.
by calling hostapd_set_wds_sta(). Since Multi-AP requires WPA2 (never
WEP), we can safely call hostapd_set_wds_encryption() as well and we
can reuse the entire WDS condition.

To parse the Multi-AP Extension subelement, we use get_ie(): even
though that function is meant for parsing IEs, it works for subelements
as well since they also have a 1-byte type and 1-byte length.

Signed-off-by: Venkateswara Naralasetty <vnaralas@codeaurora.org>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
---
v2 (Arnout):
- Rename multi_ap_enabled -> multi_ap
- Rewrite hostapd_validate_multi_ap_ie() to make it similar to the other
  check_*() functions:
  - condition on conf is moved inside the function;
  - takes multi_ap and multi_ap_len as arguments;
  - rename to check_multi_ap();
  - properly validate the IE against the multi_ap setting.
- Remove struct multi_ap_ie since the contents is a list of attributes
  so it can't be mapped on a single struct.
  - check_multi_ap() uses get_ie() instead of casting to
    struct multi_ap_ie;
  - in hostapd_eid_multi_ap(), construct directly into eid.
- Don't check length in ieee802_11_parse_vendor_specific(), since it
  is already checked in get_ie().
- Factor out add_multi_ap_ie() so it can be used in later patches.
- Extend the commit message.