ieee802_11_common: factor out add_multi_ap_ie()
The multi-AP information element is constructed separately in
hostapd_eid_multi_ap() and wpa_add_multi_ap_info_ie(). Replace these
with a single function add_multi_ap_ie() which takes the value of the IE
as argument.

Although this does not save any lines of code since the caller still has
to do error handling, it does make the code more readable.

While we're at it, also error out when the IE doesn't fit in the buffer.
Indeed, leaving out the IE and still assuming that the link is a
backhaul link will be wrong (one side will set 4addr mode and the other
will not).

Note that proper handling in case the buffer is too small is missing in
hostapd_eid_multi_ap(), but the same is true for the rest of the file.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Cc: Venkateswara Naralasetty <vnaralas@codeaurora.org>