wpa_supplicant: SME: correct use of multi_ap_backhaul_sta
In the SME case when there are multiple SSIDs configured, we should use
the multi_ap_backhaul_sta setting of the current SSID, not arbitrarily
the first one.

Also, in general (e.g. when performing WPS) there may be no configured
SSID at all.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Cc: Venkateswara Naralasetty <vnaralas@codeaurora.org>
---
I'm not 100% sure about this patch. But I tested it (with a single SSID)
and it still works. And it looks obvious to me that using the first
configured SSID makes no sense.